Add 'latest' filter to AnalysisJob viewset
Moves the existing filterset class for AnalysisJob into the pfb_analysis
module and adds filtering for 'latest', which filters the list to only
those jobs that are the latest job created for their neighborhood.
