+++ src/django/pfb_analysis/views.py
@@ -4,6 +4,7 @@ from datetime import datetime
 import us
 
 from django.utils.text import slugify
+from django.db.models import Case, When, Q, Max, Value, BooleanField
 
 from rest_framework import status
 from rest_framework.decorators import detail_route
@@ -12,13 +13,13 @@ from rest_framework.views import APIView
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.response import Response
 
-from pfb_analysis.models import AnalysisJob, Neighborhood
-from pfb_analysis.serializers import AnalysisJobSerializer, NeighborhoodSerializer
 from pfb_network_connectivity.pagination import OptionalLimitOffsetPagination
-from pfb_network_connectivity.filters import (OrgAutoFilterBackend,
-                                              SelfUserAutoFilterBackend,
-                                              AnalysisJobStatusFilterSet)
+from pfb_network_connectivity.filters import OrgAutoFilterBackend
 from pfb_network_connectivity.permissions import IsAdminOrgAndAdminCreateEditOnly, RestrictedCreate
+
+from .models import AnalysisJob, Neighborhood
+from .serializers import AnalysisJobSerializer, NeighborhoodSerializer
+from .filters import AnalysisJobFilterSet
 
 
 class AnalysisJobViewSet(ModelViewSet):
@@ -27,7 +28,7 @@ class AnalysisJobViewSet(ModelViewSet):
     queryset = AnalysisJob.objects.all()
     serializer_class = AnalysisJobSerializer
     permission_classes = (RestrictedCreate,)
-    filter_class = AnalysisJobStatusFilterSet
+    filter_class = AnalysisJobFilterSet
     filter_backends = (DjangoFilterBackend, OrderingFilter, OrgAutoFilterBackend)
     ordering_fields = ('created_at', 'modified_at')
     ordering = ('-created_at',)
